feat: Extend mocks and IAVLDatabase and tests for changes in Database interface
- Extend the MockMutableTree mock to include new methods:
  - WorkingVersion
  - LoadVersion
  - WorkingHash
  - AvailableVersions
  - IsEmpty
- Extend the MockDatabase mock to include new methods:
  - WorkingVersion
  - SaveVersion
  - Load
  - LoadVersion
  - String
  - WorkingHash
  - AvailableVersions
  - IsEmpty
- Implement new methods in IAVLDatabase:
  - WorkingVersion
  - SaveVersion
  - Load
  - LoadVersion
  - String
  - WorkingHash
  - AvailableVersions
  - IsEmpty

This commit introduces a new feature by extending the mock implementations and IAVLDatabase to accommodate changes made in the Database interface.

+++ nova/pkg/database/database.go
@@ -57,7 +57,7 @@ type Database interface {
 	WorkingHash() []byte
 
 	// AvailableVersions returns a list of available versions.
-	AvailableVersions() []int64
+	AvailableVersions() []int
 
 	// IsEmpty checks if the database is empty.
 	IsEmpty() bool
